requirements.txt
streamlit==1.32.0
ultralytics==8.2.0
opencv-python-headless==4.7.0.72
numpy==1.24.4
Pillow